% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.r
\name{cap_bib_field}
\alias{cap_bib_field}
\title{Capitalizing the first letter of a selected field entry in a Bibtex file}
\usage{
cap_bib_field(file_path, field, exceptions = NULL)
}
\arguments{
\item{file_path}{A file_path where a Bibtex file is located.r}

\item{field}{A selected field (e.g., "title")}

\item{exceptions}{A character vector that contains the words to avoid capitalization.}
}
\value{
A Bibtex object where the selected filed is modified.
}
\description{
Capitalizing the first letter of a selected field entry in a Bibtex file
}
