% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.r
\name{make_exceptions}
\alias{make_exceptions}
\title{Make exceptions for capitalizing the first letter of a string vector.}
\usage{
make_exceptions(word, replacement)
}
\arguments{
\item{replacement}{The target string object that contains the target word.}

\item{pattern}{The target word to avoid capitalization.}
}
\value{
The target string object in which the target word becomes lower cased.
}
\description{
Make exceptions for capitalizing the first letter of a string vector.
}
